adding tests of interactive filtering; splitting reconstruction tests according to SPA/ET
